chat feature added

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -56,16 +56,46 @@ export async function POST(
     if (!loggedInUser) {
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
-    await prisma.like.upsert({
+
+    const post = await prisma.post.findUnique({
       where: {
-        userId_postId: {
-          userId: loggedInUser.id,
-          postId: postId,
-        },
+        id: postId,
       },
-      create: { userId: loggedInUser.id, postId: postId },
-      update: {},
+      select: {
+        userId: true,
+      },
     });
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    await prisma.$transaction([
+      prisma.like.upsert({
+        where: {
+          userId_postId: {
+            userId: loggedInUser.id,
+            postId: postId,
+          },
+        },
+        create: { userId: loggedInUser.id, postId: postId },
+        update: {},
+      }),
+      //Notifications
+      ...(loggedInUser.id !== post.userId
+        ? [
+            prisma.notification.create({
+              data: {
+                issuerId: loggedInUser.id,
+                recipientId: post.userId,
+                type: "LIKE",
+                postId: postId,
+              },
+            }),
+          ]
+        : []),
+    ]);
+
     return new Response();
   } catch (err) {
     console.log(err);
@@ -83,12 +113,37 @@ export async function DELETE(
     if (!loggedInUser) {
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
-    await prisma.like.deleteMany({
+
+    const post = await prisma.post.findUnique({
       where: {
-        userId: loggedInUser.id,
-        postId: postId,
+        id: postId,
+      },
+      select: {
+        userId: true,
       },
     });
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    await prisma.$transaction([
+      prisma.like.deleteMany({
+        where: {
+          userId: loggedInUser.id,
+          postId: postId,
+        },
+      }),
+      prisma.notification.deleteMany({
+        where: {
+          issuerId: loggedInUser.id,
+          recipientId: post.userId,
+          type: "LIKE",
+          postId: postId,
+        },
+      }),
+    ]);
+
     return new Response();
   } catch (err) {
     console.log(err);
